sheets_automation/requirements.txt
# Google Sheets API dependencies for OAuth
google-api-python-client==2.105.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1